requirements.txt
streamlit
PyPDF2
pdf2image
pytesseract
Pillow
